requirements.txt
google-adk==1.27.2
python-dotenv>=1.0.0
gunicorn>=22.0.0
